Rename usecases packages
