+++ app/src/main/java/com/maximapps/maxim_weather/mainScreen/usecases/fetchforecastbycoordinates/FetchForecastByCoordinatesImpl.kt
@@ -6,14 +6,14 @@ import com.maximapps.maxim_weather.mainScreen.usecases.common.WeatherData
 import io.reactivex.rxjava3.core.Single
 import javax.inject.Inject
 
-typealias FetchForecastByCoordinatesUseCase = SingleUseCase<Nothing, WeatherData>
+typealias FetchForecastByCoordinates = SingleUseCase<Nothing, WeatherData>
 
-const val FetchForecastByCoordinates = "fetch_forecast_by_coordinates"
+const val FETCH_FORECAST_BY_COORDINATES = "fetch_forecast_by_coordinates"
 
 class FetchForecastByCoordinatesImpl @Inject constructor(
     private val locationRepository: LocationRepository,
     private val locationWeatherRepository: LocationWeatherRepository
-) : FetchForecastByCoordinatesUseCase {
+) : FetchForecastByCoordinates {
 
     @RequiresPermission("android.permission.ACCESS_FINE_LOCATION")
     override operator fun invoke(params: Nothing?): Single<WeatherData> = locationRepository
