+++ app/src/main/java/com/maximapps/maxim_weather/mainScreen/usecases/fetchforecastbyname/FetchForecastByNameImpl.kt
@@ -1,11 +1,13 @@
 package com.maximapps.maxim_weather.mainScreen.usecases.fetchforecastbyname
 
 import com.maximapps.maxim_weather.common.SingleUseCase
-import com.maximapps.maxim_weather.mainScreen.usecases.models.WeatherData
+import com.maximapps.maxim_weather.mainScreen.usecases.common.WeatherData
 import io.reactivex.rxjava3.core.Single
 import javax.inject.Inject
 
 typealias FetchForecastByName = SingleUseCase<FetchForecastByNameImpl.Params, WeatherData>
+
+const val FETCH_FORECAST_BY_NAME = "fetch_forecast_by_name"
 
 class FetchForecastByNameImpl @Inject constructor(
     private val cityWeatherRepository: CityWeatherRepository
